Rename system database from starrocks to sys
