+++ fe/fe-core/src/main/java/com/starrocks/catalog/system/SystemId.java
@@ -100,7 +100,7 @@ public class SystemId {
 
     public static final long STREAM_LOADS_ID = 41L;
 
-    public static final long STARROCKS_DB_ID = 100L;
+    public static final long SYS_DB_ID = 100L;
 
     public static final long ROLE_EDGES_ID = 101L;
     public static final long GRANTS_TO_ROLES_ID = 102L;
