+++ fe/fe-core/src/main/java/com/starrocks/catalog/system/sys/RoleEdges.java
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package com.starrocks.catalog.system.starrocks;
+package com.starrocks.catalog.system.sys;
 
 import com.starrocks.catalog.ScalarType;
 import com.starrocks.catalog.Table;
