+++ fe/fe-core/src/main/java/com/starrocks/service/FrontendServiceImpl.java
@@ -63,8 +63,8 @@ import com.starrocks.catalog.Table;
 import com.starrocks.catalog.Table.TableType;
 import com.starrocks.catalog.Tablet;
 import com.starrocks.catalog.View;
-import com.starrocks.catalog.system.starrocks.GrantsTo;
-import com.starrocks.catalog.system.starrocks.RoleEdges;
+import com.starrocks.catalog.system.sys.GrantsTo;
+import com.starrocks.catalog.system.sys.RoleEdges;
 import com.starrocks.cluster.ClusterNamespace;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.common.AuthenticationException;
