+++ fe/fe-core/src/main/java/com/starrocks/sql/analyzer/DropStmtAnalyzer.java
@@ -24,7 +24,7 @@ import com.starrocks.catalog.Table;
 import com.starrocks.catalog.View;
 import com.starrocks.catalog.system.SystemId;
 import com.starrocks.catalog.system.information.InfoSchemaDb;
-import com.starrocks.catalog.system.starrocks.StarRocksDb;
+import com.starrocks.catalog.system.sys.SysDb;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.common.ErrorCode;
 import com.starrocks.common.ErrorReport;
@@ -126,9 +126,9 @@ public class DropStmtAnalyzer {
             String dbName = statement.getDbName();
             if (dbName.equalsIgnoreCase(InfoSchemaDb.DATABASE_NAME)) {
                 ErrorReport.reportSemanticException(ErrorCode.ERR_DB_ACCESS_DENIED, context.getQualifiedUser(), dbName);
-            } else if (dbName.equalsIgnoreCase(StarRocksDb.DATABASE_NAME)) {
-                Database db = GlobalStateMgr.getCurrentState().getDb(StarRocksDb.DATABASE_NAME.toLowerCase());
-                if (db.getId() == SystemId.STARROCKS_DB_ID) {
+            } else if (dbName.equalsIgnoreCase(SysDb.DATABASE_NAME)) {
+                Database db = GlobalStateMgr.getCurrentState().getDb(SysDb.DATABASE_NAME.toLowerCase());
+                if (db.getId() == SystemId.SYS_DB_ID) {
                     ErrorReport.reportSemanticException(ErrorCode.ERR_DB_ACCESS_DENIED, context.getQualifiedUser(), dbName);
                 }
             }
