+++ test/sql/test_rbac/R/test_grants_to_system_table
@@ -8,7 +8,7 @@ create user IF NOT EXISTS u1;
 grant create table on database test_grant_to_system_table to u1;
 -- result:
 -- !result
-select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from starrocks.grants_to_users where GRANTEE="'u1'@'%'" and OBJECT_DATABASE="test_grant_to_system_table";
+select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from sys.grants_to_users where GRANTEE="'u1'@'%'" and OBJECT_DATABASE="test_grant_to_system_table";
 -- result:
 'u1'@'%'	DATABASE	CREATE TABLE
 -- !result
@@ -29,7 +29,7 @@ create user IF NOT EXISTS u2;
 grant select on table t1 to u2;
 -- result:
 -- !result
-select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from starrocks.grants_to_users where GRANTEE="'u2'@'%'" and OBJECT_DATABASE=database() and OBJECT_NAME="t1" and PRIVILEGE_TYPE="SELECT";
+select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from sys.grants_to_users where GRANTEE="'u2'@'%'" and OBJECT_DATABASE=database() and OBJECT_NAME="t1" and PRIVILEGE_TYPE="SELECT";
 -- result:
 'u2'@'%'	TABLE	SELECT
 -- !result
@@ -53,7 +53,7 @@ create user IF NOT EXISTS u3;
 grant select on view v to u3;
 -- result:
 -- !result
-select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from starrocks.grants_to_users where GRANTEE="'u3'@'%'" and OBJECT_DATABASE=database() and OBJECT_NAME="v" and PRIVILEGE_TYPE="SELECT";
+select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from sys.grants_to_users where GRANTEE="'u3'@'%'" and OBJECT_DATABASE=database() and OBJECT_NAME="v" and PRIVILEGE_TYPE="SELECT";
 -- result:
 'u3'@'%'	VIEW	SELECT
 -- !result
@@ -77,7 +77,7 @@ create user IF NOT EXISTS u4;
 grant select on materialized view mv to u4;
 -- result:
 -- !result
-select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from starrocks.grants_to_users where GRANTEE="'u4'@'%'" and OBJECT_DATABASE=database() and OBJECT_NAME="mv" and PRIVILEGE_TYPE="SELECT";
+select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from sys.grants_to_users where GRANTEE="'u4'@'%'" and OBJECT_DATABASE=database() and OBJECT_NAME="mv" and PRIVILEGE_TYPE="SELECT";
 -- result:
 'u4'@'%'	MATERIALIZED VIEW	SELECT
 -- !result
