+++ test/sql/test_rbac/T/test_grants_to_system_table
@@ -2,7 +2,7 @@
 create database test_grant_to_system_table;
 create user IF NOT EXISTS u1;
 grant create table on database test_grant_to_system_table to u1;
-select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from starrocks.grants_to_users where GRANTEE="'u1'@'%'" and OBJECT_DATABASE="test_grant_to_system_table";
+select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from sys.grants_to_users where GRANTEE="'u1'@'%'" and OBJECT_DATABASE="test_grant_to_system_table";
 
 -- name: test_grant_table
 CREATE TABLE `t1` (
@@ -15,7 +15,7 @@ PROPERTIES (
 );
 create user IF NOT EXISTS u2;
 grant select on table t1 to u2;
-select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from starrocks.grants_to_users where GRANTEE="'u2'@'%'" and OBJECT_DATABASE=database() and OBJECT_NAME="t1" and PRIVILEGE_TYPE="SELECT";
+select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from sys.grants_to_users where GRANTEE="'u2'@'%'" and OBJECT_DATABASE=database() and OBJECT_NAME="t1" and PRIVILEGE_TYPE="SELECT";
 
 -- name: test_grant_view
 CREATE TABLE `t1` (
@@ -29,7 +29,7 @@ PROPERTIES (
 create view v as select * from t1;
 create user IF NOT EXISTS u3;
 grant select on view v to u3;
-select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from starrocks.grants_to_users where GRANTEE="'u3'@'%'" and OBJECT_DATABASE=database() and OBJECT_NAME="v" and PRIVILEGE_TYPE="SELECT";
+select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from sys.grants_to_users where GRANTEE="'u3'@'%'" and OBJECT_DATABASE=database() and OBJECT_NAME="v" and PRIVILEGE_TYPE="SELECT";
 
 -- name: test_grant_materialized_view
 CREATE TABLE `t1` (
@@ -43,4 +43,4 @@ PROPERTIES (
 create materialized view mv DISTRIBUTED BY HASH(`v1`) BUCKETS 12 REFRESH ASYNC as select v1,v2,v3 from t1;
 create user IF NOT EXISTS u4;
 grant select on materialized view mv to u4;
-select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from starrocks.grants_to_users where GRANTEE="'u4'@'%'" and OBJECT_DATABASE=database() and OBJECT_NAME="mv" and PRIVILEGE_TYPE="SELECT";
+select GRANTEE, OBJECT_TYPE, PRIVILEGE_TYPE from sys.grants_to_users where GRANTEE="'u4'@'%'" and OBJECT_DATABASE=database() and OBJECT_NAME="mv" and PRIVILEGE_TYPE="SELECT";
